pyproject.toml
[tool.poetry]
name = "backend-habits"
version = "0.1.0"
description = ""
authors = ["egor <sharapov.egor00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
ipython = "^8.25.0"
djangorestframework-simplejwt = "^5.3.1"
aiogram = "^3.8.0"

[tool.poetry.group.lint.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"
flake8 = "^6.1.0"

[tool.black]
# максимальная длина строки
line-length = 119
# файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # исключить несколько общих каталогов в
    | \.git          # корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # также отдельно исключить файл с именем foo.py в
                     # корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
